# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Autonomous Recruiting Agent Swarm",
    "description": "AI-powered multi-agent system for recruiting",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/analysis/repo": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyze Repo",
        "description": "Analyze a GitHub repository using LLM.",
        "operationId": "analyze_repo_api_analysis_repo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepoAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepoAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RepoAnalysisRequest": {
        "properties": {
          "repo_url": {
            "type": "string",
            "title": "Repo Url"
          }
        },
        "type": "object",
        "required": [
          "repo_url"
        ],
        "title": "RepoAnalysisRequest"
      },
      "RepoAnalysisResponse": {
        "properties": {
          "analysis": {
            "type": "string",
            "title": "Analysis"
          }
        },
        "type": "object",
        "required": [
          "analysis"
        ],
        "title": "RepoAnalysisResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}